Undo function relocation
