+++ extensions/amp-story/1.0/utils.js
@@ -323,22 +323,3 @@ export function triggerClickFromLightDom(anchorElement, domElement) {
   outerAnchor.click();
   outerAnchor.remove();
 }
-
-/**
- * Makes a proxy URL if document is served from a proxy origin. No-op otherwise.
- * @param {string} url
- * @param {!Element} pageEl
- * @return {string}
- */
-export const maybeMakeProxyUrl = (url, pageEl) => {
-  const urlService = Services.urlForDoc(pageEl);
-  const loc = pageEl.getAmpDoc().win.location;
-  if (!urlService.isProxyOrigin(loc.origin) || urlService.isProxyOrigin(url)) {
-    return url;
-  }
-  const resolvedRelativeUrl = urlService.resolveRelativeUrl(
-    url,
-    urlService.getSourceOrigin(loc.href)
-  );
-  return loc.origin + '/i/s/' + resolvedRelativeUrl.replace(/https?:\/\//, '');
-};
